move project functions to their own file
