+++ client/features/project/project.loadable.ts
@@ -4,6 +4,6 @@ export interface ILoadable {
   loading: string;
 }
 
-export class LoadableProject extends Project implements ILoadable  {
+export default class LoadableProject extends Project implements ILoadable  {
   loading: string;
 }
